requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.1

pyserial>=3.2

requests


